Reintroduce broken point calculation for weeks one and two
Points earned in a week were calculated in both getUserPoint and
getCouponsRemaining, but in different ways:

  * getCouponsRemaining would correctly convert each workout to
    points, counting half a point for each complete quarter-hour
    and dropping any leftovers,

  * but getUserPoint would incorrectly add up the duration of all
    workouts and convert that total to points, thereby consolidating
    leftovers and possibly giving the user extra points.

Consequently, points and coupons were sometimes inconsistent.  For
example, in a week with 4pts goal and two workouts of 55min and
67min:

  * in getUserPoint:
    * total duration: 55min + 67min = 122min
    * points earned: Math.trunc(122 / 15) / 2 = 4.0

  * in getCouponsRemaining:
    * workout #1: 55min, 1.5pts  # Math.trunc(55 / 15) / 2
    * workout #2: 67min, 2.0pts  # Math.truns(67 / 15) / 2
    * points earned: 1.5 + 2.0 = 3.5
    * points challenge: 4.0
    * coupons consumed: 0.5

Fixing getUserPoint eliminates the inconsistency but changes point
calculation retroactively for the first two weeks, surprising and
possibly upsetting users who find themselves with fewer points than
before.

The new flag KEEP_BROKEN_POINT_CALCULATION_FOR_EARLY_WEEKS in the
SurvivorWeek object, now set to true, causes the calculation of
both points and coupons for weeks one and two to add up the duration
of all workouts and convert that total to points, consolidating
leftover minutes.  This is unfair to users of the paper log, whose
points are calculated according to the proper rule, but arguably
fair to users of the app who have already been granted the extra
points.

Set KEEP_BROKEN_POINT_CALCULATION_FOR_EARLY_WEEKS to false in data
to revert this change.